+++ docs/rest-api/endpoints/examples/content/export_inline.ts
@@ -1,0 +1,14 @@
+import { Client } from '@plone/nick';
+
+const cli = Client.initialize({ apiPath: 'http://localhost:8080' });
+const login = await cli.login({
+  data: { login: 'admin', password: 'admin' },
+});
+
+const { data } = await cli.exportContent({
+  token: login.data.token,
+  path: '/news',
+  query: {
+    disposition: 'inline',
+  },
+});